requirements.txt
streamlit==1.33.0
streamlit_navigation_bar
face-recognition==1.3.0
gspread==5.11.0
oauth2client==4.1.3
opencv-python==4.8.0.74
numpy==1.23.5
pandas==2.1.1
Pillow==10.0.1
mtcnn==0.1.1
python-dotenv
